# contracts/contract.ts
// Generated by scripts/generate_frontend_contract.py; do not edit.
export type Activity = "low" | "light" | "moderate" | "high";
export type Answers = {
  sex: Sex | null;
  goal: Goal | null;
  age: number | null;
  heightCm: number | null;
  weightKg: number | null;
  targetWeightKg: number | null;
  activity: Activity | null;
  secondaryGoals: (SecondaryGoal)[] | null;
  experience: Experience | null;
  dailyActivity: DailyActivity | null;
  limitations: (Limitation)[] | null;
  sleep: Sleep | null;
  energy: Energy | null;
  mealRhythm: MealRhythm | null;
  foodHabits: (FoodHabit)[] | null;
  barrier: Barrier | null;
  timeWindow: TimeWindow | null;
};
export type AssessmentStatus = "draft" | "completed";
export type AssessmentView = {
  planPreview: PlanPreview | null;
  id: string;
  status: AssessmentStatus;
  version: number;
  flowVersion: string;
  resumeStepId: Step;
  answers: Answers;
  completedSteps: (string)[];
  missingFields: (string)[];
  updatedAt: string;
};
export type Barrier = "time" | "motivation" | "unsure" | "no_barrier";
export type BmiCategory = "underweight" | "reference" | "overweight" | "high";
export type Calculation = {
  planPreview: PlanPreview | null;
  guidance: Guidance | null;
  algorithmVersion: RuleVersion;
  calculatedOn: string;
  bmi: number;
  bmiCategory: BmiCategory;
  restingKcal: number;
  maintenanceKcal: number;
  suggestedKcal: number;
  predictedGoalDate: string;
  weeklyChangeKg: number;
  projection: (ProjectionPoint)[];
  assumptions: (CalculationAssumption)[];
};
export type CalculationAssumption = "legacy_energy" | "resting_energy" | "demo_energy_budget" | "linear_projection" | "adult_scope";
export type Currency = "CNY";
export type DailyActivity = "seated" | "mixed" | "on_feet";
export type Energy = "low_energy" | "afternoon_dip" | "steady";
export type Experience = "beginner" | "returning" | "regular";
export type FoodHabit = "late_snacks" | "sweet_drinks" | "sweets" | "none";
export type FreeResult = {
  planPreview: PlanPreview | null;
  access: "free";
  assessmentId: string;
  bmi: number;
  bmiCategory: BmiCategory;
  summary: ResultSummary;
  lockedFeatures: (string)[];
};
export type Goal = "lose" | "maintain" | "gain";
export type Guidance = {
  revision: number;
  status: GuidanceStatus;
  rulesVersion: RuleVersion;
  model: string | null;
  focus: PlanFocus | null;
  confidence: number | null;
  effortScore: number | null;
  effortConfidence: number | null;
  failureCode: string | null;
};
export type GuidanceStatus = "ready" | "uncertain" | "unavailable" | "disabled";
export type Limitation = "back" | "knees" | "none";
export type MealRhythm = "regular_meals" | "skipped_meals" | "irregular_meals";
export type MemberResult = {
  planPreview: PlanPreview | null;
  access: "member";
  assessmentId: string;
  bmi: number;
  bmiCategory: BmiCategory;
  summary: ResultSummary;
  calculation: Calculation;
};
export type NextAction = "wait" | "open_checkout" | "none";
export type PaymentStatus = "pending" | "succeeded" | "failed" | "closed";
export type PaymentView = {
  id: string;
  planId: PlanId;
  status: PaymentStatus;
  amountMinor: number;
  currency: Currency;
  checkoutUrl: string | null;
  nextAction: NextAction;
  simulated: true;
  createdAt: string;
  updatedAt: string;
};
export type PlanCard = {
  topic: PlanTopic;
  title: PlanMessage;
  reason: (PlanMessage)[];
  action: (PlanMessage)[];
  sourceFields: (string)[];
};
export type PlanCode = "movement_title" | "recovery_title" | "meals_title" | "consistency_title" | "experience" | "activity" | "seated" | "on_feet" | "movement_break" | "observe_activity" | "familiar_activity" | "gentle_movement" | "familiar_strength" | "seek_advice" | "limitation_boundary" | "sleep_energy" | "record_sleep" | "keep_sleep" | "meal_habits" | "protect_meal" | "unsweetened_drink" | "record_snack" | "record_sweets" | "keep_meals" | "barrier" | "pair_action" | "record_action" | "try_easiest" | "review_routine" | "schedule_action" | "preference";
export type PlanFocus = "movement_habit" | "meal_routine" | "consistency" | "insufficient_context";
export type PlanId = "wellnest-demo";
export type PlanMessage = {
  code: PlanCode;
  params: PlanParams;
};
export type PlanParams = {
  experience: Experience | null;
  activity: Activity | null;
  sleep: Sleep | null;
  energy: Energy | null;
  mealRhythm: MealRhythm | null;
  foodHabits: (FoodHabit)[] | null;
  barrier: Barrier | null;
  timeWindow: TimeWindow | null;
  limitations: (Limitation)[] | null;
  preference: SecondaryGoal | null;
};
export type PlanPreview = {
  formatVersion: 2;
  rulesVersion: PlanRuleVersion;
  cards: (PlanCard)[];
  preferences: (PlanMessage)[];
  boundaries: (PlanMessage)[];
};
export type PlanRuleVersion = "lifestyle-plan-v1";
export type PlanTopic = "movement" | "recovery" | "meals" | "consistency";
export type ProjectionPoint = {
  date: string;
  weightKg: number;
};
export type ResultSummary = "personal_start";
export type RuleVersion = "wellness-v1" | "wellness-v2" | "plan-guidance-v1" | "plan-guidance-v2" | "plan-guidance-v3";
export type SecondaryGoal = "energy" | "mobility" | "routine" | "strength";
export type SessionView = {
  sessionId: string;
  subscriptionStatus: SubscriptionStatus;
  assessment: AssessmentView | null;
};
export type Sex = "female" | "male";
export type Sleep = "short" | "variable" | "rested";
export type Step = "welcome" | "sex" | "goal" | "age" | "height" | "weight" | "target" | "activity" | "review" | "result" | "goal_feedback" | "activity_feedback" | "target_feedback" | "secondary_goals" | "experience" | "daily_activity" | "limitations" | "sleep" | "energy" | "meal_rhythm" | "food_habits" | "barrier" | "time_window" | "movement_feedback" | "habits_feedback";
export type SubscriptionStatus = "inactive" | "active";
export type TimeWindow = "morning" | "midday" | "evening" | "varies";
export const inputRules = {"ageMin": 18, "ageMax": 80, "heightMinCm": 120, "heightMaxCm": 220, "weightMinKg": 35, "weightMaxKg": 250, "maxTargetChangeRatio": 0.25, "targetBmiMin": 18.5, "targetBmiMax": 40} as const;
export const steps = ["welcome", "sex", "goal", "age", "height", "weight", "target", "activity", "review"] as const;
export const guidedFlowVersion = "guided-v2" as const;
export const lifestyleFlowVersion = "lifestyle-v3" as const;
export const flowSteps = {"wellness-v1": ["welcome", "sex", "goal", "age", "height", "weight", "target", "activity", "review"], "guided-v2": ["welcome", "goal", "goal_feedback", "sex", "activity", "activity_feedback", "age", "height", "weight", "target", "target_feedback", "review"], "lifestyle-v3": ["welcome", "goal", "goal_feedback", "sex", "secondary_goals", "experience", "activity", "daily_activity", "limitations", "movement_feedback", "sleep", "energy", "meal_rhythm", "food_habits", "barrier", "time_window", "habits_feedback", "age", "height", "weight", "target", "target_feedback", "review"]} as const;
export const fields = {"sex": "sex", "goal": "goal", "age": "age", "height": "heightCm", "weight": "weightKg", "target": "targetWeightKg", "activity": "activity", "secondary_goals": "secondaryGoals", "experience": "experience", "daily_activity": "dailyActivity", "limitations": "limitations", "sleep": "sleep", "energy": "energy", "meal_rhythm": "mealRhythm", "food_habits": "foodHabits", "barrier": "barrier", "time_window": "timeWindow"} as const;
export const cmPerMetre = 100 as const;
export const gradualEffortThreshold = 1 as const;