# contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Wellnest API",
    "version": "0.2.0"
  },
  "paths": {
...
    "/api/payments": {
      "post": {
        "tags": [
          "payments"
        ],
        "summary": "Create Payment",
        "operationId": "create_payment_api_payments_post",
        "parameters": [
          {
            "name": "idempotency-key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 8,
              "maxLength": 128,
              "title": "Idempotency-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PayInput"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaymentView"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/payments/{payment_id}": {
      "get": {
        "tags": [
          "payments"
        ],
        "summary": "Get Payment",
        "operationId": "get_payment_api_payments__payment_id__get",
        "parameters": [
          {
            "name": "payment_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Payment Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaymentView"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/payments/{payment_id}/refresh": {
      "post": {
        "tags": [
          "payments"
        ],
        "summary": "Refresh",
        "operationId": "refresh_api_payments__payment_id__refresh_post",
        "parameters": [
          {
            "name": "payment_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Payment Id"
            }
          }
        ],
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Accepted"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/webhooks/payments/mock": {
      "post": {
        "tags": [
          "payments"
        ],
        "summary": "Webhook",
        "operationId": "webhook_api_webhooks_payments_mock_post",
        "parameters": [
          {
            "name": "x-payment-timestamp",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "X-Payment-Timestamp"
            }
          },
          {
            "name": "x-payment-signature",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "X-Payment-Signature"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Accepted"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": false,
                "properties": {
                  "merchant_payment_id": {
                    "format": "uuid",
                    "title": "Merchant Payment Id",
                    "type": "string"
                  },
                  "amount_minor": {
                    "exclusiveMinimum": 0,
                    "title": "Amount Minor",
                    "type": "integer"
                  },
                  "currency": {
                    "$ref": "#/components/schemas/Currency"
                  },
                  "transaction_id": {
                    "format": "uuid",
                    "title": "Transaction Id",
                    "type": "string"
                  },
                  "status": {
                    "$ref": "#/components/schemas/PaymentStatus"
                  },
                  "checkout_url": {
                    "title": "Checkout Url",
                    "type": "string"
                  },
                  "expires_at": {
                    "format": "date-time",
                    "title": "Expires At",
                    "type": "string"
                  },
                  "event_id": {
                    "format": "uuid",
                    "title": "Event Id",
                    "type": "string"
                  }
                },
                "required": [
                  "merchant_payment_id",
                  "amount_minor",
                  "currency",
                  "transaction_id",
                  "status",
                  "checkout_url",
                  "expires_at",
                  "event_id"
                ],
                "title": "ChannelNotification",
                "type": "object"
              }
            }
          }
        }
      }
    },
    "/api/mock-provider/payments": {
      "post": {
        "tags": [
          "payments"
        ],
        "summary": "Provider Create",
        "operationId": "provider_create_api_mock_provider_payments_post",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChannelCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChannelPayment"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/mock-provider/payments/{merchant_payment_id}": {
      "get": {
        "tags": [
          "payments"
        ],
        "summary": "Provider Query",
        "operationId": "provider_query_api_mock_provider_payments__merchant_payment_id__get",
        "parameters": [
          {
            "name": "merchant_payment_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Merchant Payment Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChannelPayment"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/mock-checkout/{token}": {
      "get": {
        "tags": [
          "payments"
        ],
        "summary": "Checkout",
        "operationId": "checkout_api_mock_checkout__token__get",
        "parameters": [
          {
            "name": "token",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Token"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChannelPayment"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/mock-checkout/{token}/confirm": {
      "post": {
        "tags": [
          "payments"
        ],
        "summary": "Confirm",
        "operationId": "confirm_api_mock_checkout__token__confirm_post",
        "parameters": [
          {
            "name": "token",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Token"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CheckoutConfirm"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChannelPayment"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Accepted": {
        "properties": {
          "accepted": {
            "type": "boolean",
            "const": true,
            "title": "Accepted",
            "default": true
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "Accepted"
      },
...
      "ChannelCreate": {
        "properties": {
          "merchant_payment_id": {
            "type": "string",
            "format": "uuid",
            "title": "Merchant Payment Id"
          },
          "amount_minor": {
            "type": "integer",
            "exclusiveMinimum": 0.0,
            "title": "Amount Minor"
          },
          "currency": {
            "$ref": "#/components/schemas/Currency"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "merchant_payment_id",
          "amount_minor",
          "currency"
        ],
        "title": "ChannelCreate"
      },
      "ChannelPayment": {
        "properties": {
          "merchant_payment_id": {
            "type": "string",
            "format": "uuid",
            "title": "Merchant Payment Id"
          },
          "amount_minor": {
            "type": "integer",
            "exclusiveMinimum": 0.0,
            "title": "Amount Minor"
          },
          "currency": {
            "$ref": "#/components/schemas/Currency"
          },
          "transaction_id": {
            "type": "string",
            "format": "uuid",
            "title": "Transaction Id"
          },
          "status": {
            "$ref": "#/components/schemas/PaymentStatus"
          },
          "checkout_url": {
            "type": "string",
            "title": "Checkout Url"
          },
          "expires_at": {
            "type": "string",
            "format": "date-time",
            "title": "Expires At"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "merchant_payment_id",
          "amount_minor",
          "currency",
          "transaction_id",
          "status",
          "checkout_url",
          "expires_at"
        ],
        "title": "ChannelPayment"
      },
      "CheckoutConfirm": {
        "properties": {
          "outcome": {
            "type": "string",
            "enum": [
              "succeeded",
              "failed"
            ],
            "title": "Outcome",
            "default": "succeeded"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "CheckoutConfirm"
      },
...
      "Currency": {
        "type": "string",
        "enum": [
          "CNY"
        ],
        "title": "Currency"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "NextAction": {
        "type": "string",
        "enum": [
          "wait",
          "open_checkout",
          "none"
        ],
        "title": "NextAction"
      },
      "PayInput": {
        "properties": {
          "planId": {
            "$ref": "#/components/schemas/PlanId",
            "default": "wellnest-demo"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "PayInput"
      },
      "PaymentStatus": {
        "type": "string",
        "enum": [
          "pending",
          "succeeded",
          "failed",
          "closed"
        ],
        "title": "PaymentStatus"
      },
      "PaymentView": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "planId": {
            "$ref": "#/components/schemas/PlanId"
          },
          "status": {
            "$ref": "#/components/schemas/PaymentStatus"
          },
          "amountMinor": {
            "type": "integer",
            "title": "Amountminor"
          },
          "currency": {
            "$ref": "#/components/schemas/Currency"
          },
          "checkoutUrl": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Checkouturl"
          },
          "nextAction": {
            "$ref": "#/components/schemas/NextAction"
          },
          "simulated": {
            "type": "boolean",
            "const": true,
            "title": "Simulated",
            "default": true
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "title": "Createdat"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time",
            "title": "Updatedat"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "planId",
          "status",
          "amountMinor",
          "currency",
          "checkoutUrl",
          "nextAction",
          "createdAt",
          "updatedAt"
        ],
        "title": "PaymentView"
      },
...
      "PlanId": {
        "type": "string",
        "enum": [
          "wellnest-demo"
        ],
        "title": "PlanId"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}